Add possibility to copy own twiddle
If a gist owned by the authenticated user is shown, a new menu item
"Copy Twiddle" allows to copy the current files into a new Twiddle. Note
that the newly created Twiddle is not yet saved.

By this, Twiddles (e.g. Ember setup for a specific version) can be used
as a starting point without having to re-setup the whole Twiddle again
and again.