"Run now" reloads the application when there are no changes
The used `{{dummy-demo-app}}` component only reloads the application if
the passed `html` property changes. If nothing changes in the underlying
app, succeeding clicks on "run now" will do nothing, as the HTML doesn't
change.

To force enable the reload of the application, this commit adds a new
`<meta name="build-at" >` tag, which' value is set to the milliseconds
since the Unix Epoch. By this, the HTML changes everytime the app is
build and now the "run now" button always reloads the application.