Don't set version for {{user-menu}} in twiddles template
The version is read from the config inside the {{user-menu}} component,
so there is no need to pass it.

Also, this is a leftover from 55e5f895f161b1b23efc4a513b7abc66baaf5f64,
where the {{user-menu}} has been added to the toolbar but there is no
`version` property set on the controller.